gotemplate: more convenient usage of template functions
It's not suitable to enter maps form a gotemplate file, so the functions
take the map as variadic arguments alternating between key and value.

On top there was no possibility to print unescaped HTML and JS which can
now be done with the plainHTML and plainJS template functions.
